Mvvm

+++ DosingApp/DosingApp/Models/Field.cs
@@ -11,6 +11,6 @@ namespace DosingApp.Models
         [Required]
         public string Name { get; set; }
         public string Code { get; set; }
-        public float? Size { get; set; }
+        public double? Size { get; set; }
     }
 }
